Move information-panel/ to ir/cartes-impact/
